refactor(queue): extract shared schema to ingest/queue.py, add input validation
- DRY: move duplicated QUEUE_SCHEMA_DDL + connect/ensure_schema to ingest/queue.py
- Security: add validate_identifier() for group_id/session_key/source
- Docs: add security considerations section to MEMORY-RUNTIME-WIRING.md
- Both ingest_trigger_done.py and run_incremental_ingest.py now import from shared module
