+++ scripts/ingest_trigger_done.py
@@ -22,6 +22,9 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Optional
 
+# Shared queue schema + helpers (single source of truth).
+sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
+from ingest.queue import connect, ensure_schema, validate_identifier  # noqa: E402
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
 DEFAULT_DB_PATH = REPO_ROOT / "state" / "ingest_registry.db"
@@ -30,46 +33,6 @@ DEFAULT_JOB_TYPE = "sessions_incremental"
 DEFAULT_LANE = "primary"
 DEFAULT_OVERLAP_CHUNKS = 10
 DEFAULT_MAX_ATTEMPTS = 6
-
-
-QUEUE_SCHEMA_DDL = """
-CREATE TABLE IF NOT EXISTS ingest_jobs (
-  job_id TEXT PRIMARY KEY,
-  dedupe_key TEXT,
-  created_at TEXT NOT NULL,
-  updated_at TEXT NOT NULL,
-  source TEXT NOT NULL,
-  job_type TEXT NOT NULL,
-  group_id TEXT NOT NULL,
-  lane TEXT NOT NULL,
-  session_key TEXT,
-  requested_ts TEXT,
-  status TEXT NOT NULL,
-  run_after TEXT NOT NULL,
-  attempts INTEGER NOT NULL DEFAULT 0,
-  max_attempts INTEGER NOT NULL DEFAULT 6,
-  payload_json TEXT NOT NULL DEFAULT '{}',
-  last_error TEXT,
-  last_error_at TEXT,
-  last_started_at TEXT,
-  last_finished_at TEXT,
-  last_exit_code INTEGER,
-  last_duration_s REAL
-);
-
-CREATE INDEX IF NOT EXISTS idx_ingest_jobs_created_at
-  ON ingest_jobs(created_at);
-CREATE INDEX IF NOT EXISTS idx_ingest_jobs_job_type
-  ON ingest_jobs(job_type);
-CREATE INDEX IF NOT EXISTS idx_ingest_jobs_status_run_after
-  ON ingest_jobs(status, run_after);
-CREATE INDEX IF NOT EXISTS idx_ingest_jobs_dedupe_key
-  ON ingest_jobs(dedupe_key);
-
-CREATE UNIQUE INDEX IF NOT EXISTS ux_ingest_jobs_active_dedupe
-  ON ingest_jobs(dedupe_key)
-  WHERE status IN ('queued', 'running');
-"""
 
 
 def _utc_now_iso() -> str:
@@ -84,17 +47,6 @@ def _parse_iso_ts(ts: str) -> Optional[float]:
         return dt.timestamp()
     except Exception:
         return None
-
-
-def _connect(db_path: Path) -> sqlite3.Connection:
-    db_path.parent.mkdir(parents=True, exist_ok=True)
-    conn = sqlite3.connect(str(db_path), timeout=30)
-    conn.row_factory = sqlite3.Row
-    return conn
-
-
-def _ensure_queue_schema(conn: sqlite3.Connection) -> None:
-    conn.executescript(QUEUE_SCHEMA_DDL)
 
 
 def main() -> int:
@@ -116,7 +68,13 @@ def main() -> int:
     # /done jobs are sessions incremental ingest by default.
     job_type = DEFAULT_JOB_TYPE
     lane = DEFAULT_LANE
-    group_id = args.session_key
+
+    try:
+        group_id = validate_identifier(args.session_key, "session_key")
+        validate_identifier(args.source, "source")
+    except ValueError as exc:
+        print(f"validation_error: {exc}", file=sys.stderr)
+        return 1
 
     payload: dict[str, Any] = {
         "job_type": job_type,
@@ -135,9 +93,9 @@ def main() -> int:
     job_id = uuid.uuid4().hex
 
     try:
-        conn = _connect(db_path)
+        conn = connect(db_path)
         try:
-            _ensure_queue_schema(conn)
+            ensure_schema(conn)
 
             cur = conn.execute(
                 """
